Fixed memory allocation issue,
We were allocating memory to accomodate an array of vertecies the size of the mesh verts multiplied by the max allowed meshes. This was originally used for stroing multiple instances of vertex buffer data for each instance. However now with the changes to how we handle instanced rendering every uniqu mesh only need one vertex buffer array the same size as the amount of vertecies the mesh consists of. All instances of a mesh share the same vertex buffer array.